# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Motorzad API Documentation",
        "description": "API Documentation for Motorzad Car Auctions Platform.",
        "contact": {
            "email": "support@motorzad.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Motorzad/public",
            "description": "XAMPP Apache Server (Local)"
        },
        {
            "url": "http://localhost:8000",
            "description": "Artisan Development Server (php artisan serve)"
        },
        {
            "url": "/",
            "description": "Production Server (Root Domain)"
        }
    ],
    "paths": {
...
        "/api/seller-subscription/status": {
            "get": {
                "tags": [
                    "Seller"
                ],
                "summary": "Get Seller Subscription Status",
                "description": "Returns the current status of the user's seller subscription request (seller, pending, rejected, or none).",
                "operationId": "ff3431232c587f0f8152d0be325c8135",
                "responses": {
                    "200": {
                        "description": "Current status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "description": "Can be: seller, pending, rejected, none",
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "admin_notes": {
                                                    "description": "Present if status is rejected",
                                                    "type": "string",
                                                    "example": "Identity proof is not clear.",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seller-subscription/subscribe": {
            "post": {
                "tags": [
                    "Seller"
                ],
                "summary": "Submit Seller Subscription Request",
                "description": "Submits a request to upgrade the authenticated user to a seller role. Requires KYC verification to be approved. Admin approval is required.",
                "operationId": "fbdca613536660625b0b3a0905271e83",
                "responses": {
                    "200": {
                        "description": "Successfully submitted the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your request to become a seller has been submitted successfully and is awaiting admin approval."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Already a seller or KYC not approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Please complete identity verification to become a seller."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}